+++ CHANGELOG.md
@@ -13,6 +13,7 @@ FEATURES:
   - You can now specify a memory using `vb.memory` setting with VirtualBox.
   - Plugin developers can now hook into `environment_plugins_loaded`, which is
     executed after plugins are loaded but before Vagrantfiles are parsed.
+  - VirtualBox internal networks are now supported. [GH-2020]
 
 IMPROVEMENTS: